Create dir to store test data and start to move data into it
