+++ tests/data/libs/standard_sc.yml
