+++ tests/functional/test_xpansion.py
@@ -13,16 +13,8 @@
 import pandas as pd
 import pytest
 
-from andromede.expression.expression import literal, param, port_field, var
+from andromede.expression.expression import literal, param, var
 from andromede.expression.indexing_structure import IndexingStructure
-from andromede.libs.standard import (
-    BALANCE_PORT_TYPE,
-    CONSTANT,
-    DEMAND_MODEL,
-    GENERATOR_MODEL,
-    NODE_BALANCE_MODEL,
-    NODE_WITH_SPILL_AND_ENS_MODEL,
-)
 from andromede.model import (
     Constraint,
     Model,
@@ -38,7 +30,6 @@ from andromede.simulation import (
     MergedProblemStrategy,
     OutputValues,
     TimeBlock,
-    build_benders_decomposed_problem,
     build_problem,
 )
 from andromede.study import (
@@ -50,6 +41,13 @@ from andromede.study import (
     PortRef,
     TimeScenarioSeriesData,
     create_component,
+)
+from tests.data.libs.standard import (
+    BALANCE_PORT_TYPE,
+    CONSTANT,
+    DEMAND_MODEL,
+    GENERATOR_MODEL,
+    NODE_BALANCE_MODEL,
 )
 
 CONSTANT = IndexingStructure(False, False)
