+++ tests/models/test_electrolyzer_n_inputs.py
@@ -12,14 +12,6 @@
 
 import math
 
-from andromede.libs.standard import DEMAND_MODEL, GENERATOR_MODEL, NODE_BALANCE_MODEL
-from andromede.libs.standard_sc import (
-    CONVERTOR_MODEL,
-    CONVERTOR_RECEIVE_IN,
-    DECOMPOSE_1_FLOW_INTO_2_FLOW,
-    NODE_BALANCE_MODEL_MOD,
-    TWO_INPUTS_CONVERTOR_MODEL,
-)
 from andromede.simulation import OutputValues, TimeBlock, build_problem
 from andromede.study import (
     ConstantData,
@@ -28,6 +20,14 @@ from andromede.study import (
     Node,
     PortRef,
     create_component,
+)
+from tests.data.libs.standard import DEMAND_MODEL, GENERATOR_MODEL, NODE_BALANCE_MODEL
+from tests.data.libs.standard_sc import (
+    CONVERTOR_MODEL,
+    CONVERTOR_RECEIVE_IN,
+    DECOMPOSE_1_FLOW_INTO_2_FLOW,
+    NODE_BALANCE_MODEL_MOD,
+    TWO_INPUTS_CONVERTOR_MODEL,
 )
 
 """
