+++ tests/unittests/model/test_model_parsing.py
@@ -17,7 +17,6 @@ import pytest
 from andromede.expression import literal, param, var
 from andromede.expression.expression import port_field
 from andromede.expression.parsing.parse_expression import AntaresParseException
-from andromede.libs.standard import CONSTANT
 from andromede.model import (
     Constraint,
     ModelPort,
@@ -30,6 +29,7 @@ from andromede.model import (
 from andromede.model.model import PortFieldDefinition, PortFieldId
 from andromede.model.parsing import parse_yaml_library
 from andromede.model.resolve_library import resolve_library
+from tests.data.libs.standard import CONSTANT
 
 
 def test_library_parsing(data_dir: Path) -> None:
